Add ability to specify multiple days per week to run
Previously, this plugin was limited to running either every day of the
week or only one day of the week. This change allows any number of days
per week to be selected.

Days of the week are still specified by the index for that day
from Sunday (0) to Saturday (2). Values are added to the existing
run_on_dow key, but it is now a comma separated list.

For example, to run on Sunday, Wednesday and Friday, one would set
run_on_dow: 0,3,5